fix: rapi rapi fitur register
